add labels to maps
