+++ src/data/maps.json
@@ -233,7 +233,7 @@
                                 "height": [5.7, 1000],
                                 "bounds": [
                                     [[580, -104], [532, -134], "dead scav warehouse"],
-                                    [[-199, -90], [-210, -114], "big red 2nd"],
+                                    [[-199, -90], [-223, -131], "big red 2nd"],
                                     [[239, 3], [169, -160], "skeleton"],
                                     [[336, -56], [316, -95], "switch snipe"],
                                     [[584, -46], [556, -92], "USEC 2nd"],
@@ -261,6 +261,16 @@
                                 ]
                             }
                         ]
+                    }
+                ],
+                "labels": [
+                    {
+                        "position": [-212, -123],
+                        "text": "Big Red"
+                    },
+                    {
+                        "position": [202, 149],
+                        "text": "Dorms"
                     }
                 ]
             },
@@ -330,6 +340,13 @@
                                 "height": [-10000, -1]
                             }
                         ]
+                    }
+                ],
+                "labels": [
+                    {
+                        "position": [19, 41],
+                        "text": "Office",
+                        "bottom": 0
                     }
                 ]
             },
