+++ src/pages/map/map-images.mjs
@@ -1,0 +1,50 @@
+
+const images = {
+    'container_bank-cash-register': 'container_cash-register',
+    'container_bank-safe': 'container_safe',
+    'container_buried-barrel-cache': 'container_buried-barrel-cache',
+    'container_cash-register': 'container_cash-register',
+    'container_cash-register-tar2-2': 'container_cash-register',
+    'container_dead-civilian': 'container_dead-scav',
+    'container_dead-scav': 'container_dead-scav',
+    'container_pmc-body': 'container_dead-scav',
+    'container_drawer': 'container_drawer',
+    'container_duffle-bag': 'container_duffle-bag',
+    'container_grenade-box': 'container_grenade-box',
+    'container_ground-cache': 'container_buried-barrel-cache',
+    'container_jacket': 'container_jacket',
+    'container_lab-technician-body': 'container_dead-scav',
+    'container_medbag-smu06': 'container_medbag-smu06',
+    'container_medcase': 'container_medcase',
+    'container_medical-supply-crate': 'container_technical-supply-crate',
+    'container_pc-block': 'container_pc-block',
+    'container_plastic-suitcase': 'container_plastic-suitcase',
+    'container_ration-supply-crate': 'container_technical-supply-crate',
+    'container_safe': 'container_safe',
+    'container_scav-body': 'container_dead-scav',
+    'container_shturmans-stash': 'container_weapon-box',
+    'container_technical-supply-crate': 'container_technical-supply-crate',
+    'container_toolbox': 'container_toolbox',
+    'container_weapon-box': 'container_weapon-box',
+    'container_wooden-ammo-box': 'container_wooden-ammo-box',
+    'container_wooden-crate': 'container_wooden-crate',
+    'extract_pmc': 'extract_pmc',
+    'extract_scav': 'extract_scav',
+    'extract_shared': 'extract_shared',
+    'hazard': 'hazard',
+    'key': 'key',
+    'lock': 'lock',
+    'quest_item': 'quest_item',
+    'quest_objective': 'quest_objective',
+    'spawn_sniper_scav': 'spawn_sniper_scav',
+    'spawn_bloodhound': 'spawn_bloodhound',
+    'spawn_boss': 'spawn_boss',
+    'spawn_cultist-priest': 'spawn_cultist-priest',
+    'spawn_pmc': 'spawn_pmc',
+    'spawn_rogue': 'spawn_rogue',
+    'spawn_scav': 'spawn_scav',
+    'stationarygun': 'stationarygun',
+    'switch': 'switch',
+};
+
+export default images;
